Special hack for emergency patch
The handshake has been made backwards compatible (no extra word for
addressable versus unaddressable). Unaddressable is signalled by a
0-length address. This may be better than how it was before. The hackish
part is the decision to treat 0.0.0.0 and 127.0.0.1 hosts as
unaddressable.

network-transport-tests will break, because the address of the received
event will not match the endpoint's address: it doesn't choose
Unaddressable as its own address.